Move aggregation utils to main lib and update examples accordinly
