+++ examples/dp_fed_examples/instance_level_dp/server.py
@@ -6,13 +6,13 @@ import flwr as fl
 from flwr.common.typing import Config
 
 from examples.models.cnn_model import Net
-from examples.simple_metric_aggregation import evaluate_metrics_aggregation_fn, fit_metrics_aggregation_fn
 from examples.utils.functions import make_dict_with_epochs_or_steps
 from fl4health.client_managers.poisson_sampling_manager import PoissonSamplingClientManager
 from fl4health.server.instance_level_dp_server import InstanceLevelDPServer
 from fl4health.strategies.basic_fedavg import BasicFedAvg
 from fl4health.utils.config import load_config
 from fl4health.utils.functions import get_all_model_parameters
+from fl4health.utils.metric_aggregation import evaluate_metrics_aggregation_fn, fit_metrics_aggregation_fn
 
 
 def construct_config(
@@ -42,7 +42,14 @@ def fit_config(
     local_epochs: Optional[int] = None,
     local_steps: Optional[int] = None,
 ) -> Config:
-    return construct_config(current_round, batch_size, noise_multiplier, clipping_bound, local_epochs, local_steps)
+    return construct_config(
+        current_round,
+        batch_size,
+        noise_multiplier,
+        clipping_bound,
+        local_epochs,
+        local_steps,
+    )
 
 
 def main(config: Dict[str, Any]) -> None:
