+++ examples/dp_scaffold_example/server.py
@@ -8,11 +8,11 @@ from flwr.common.parameter import ndarrays_to_parameters
 from flwr.common.typing import Config, Parameters
 
 from examples.models.cnn_model import MnistNet
-from examples.simple_metric_aggregation import evaluate_metrics_aggregation_fn, fit_metrics_aggregation_fn
 from fl4health.client_managers.poisson_sampling_manager import PoissonSamplingClientManager
 from fl4health.server.scaffold_server import DPScaffoldServer
 from fl4health.strategies.scaffold import Scaffold
 from fl4health.utils.config import load_config
+from fl4health.utils.metric_aggregation import evaluate_metrics_aggregation_fn, fit_metrics_aggregation_fn
 
 
 def get_initial_model_information() -> Tuple[Parameters, Parameters]:
