+++ examples/feature_alignment_example/server.py
@@ -8,12 +8,12 @@ from flwr.common.parameter import ndarrays_to_parameters
 from flwr.common.typing import Parameters
 
 from examples.models.mlp_classifier import MLP
-from examples.simple_metric_aggregation import evaluate_metrics_aggregation_fn, fit_metrics_aggregation_fn
 from fl4health.client_managers.poisson_sampling_manager import PoissonSamplingClientManager
 from fl4health.feature_alignment.tab_features_info_encoder import TabularFeaturesInfoEncoder
 from fl4health.server.tabular_feature_alignment_server import TabularFeatureAlignmentServer
 from fl4health.strategies.basic_fedavg import BasicFedAvg
 from fl4health.utils.config import load_config
+from fl4health.utils.metric_aggregation import evaluate_metrics_aggregation_fn, fit_metrics_aggregation_fn
 
 # This data path is used to create a "source of truth" on the server-side as an example.
 # This is used if the config specifies source_specified as true
