+++ examples/federated_eval_example/server.py
@@ -4,10 +4,10 @@ from typing import Any, Dict, Optional
 
 import flwr as fl
 
-from examples.simple_metric_aggregation import uniform_evaluate_metrics_aggregation_fn
 from fl4health.client_managers.fixed_without_replacement_manager import FixedSamplingByFractionClientManager
 from fl4health.server.evaluate_server import EvaluateServer
 from fl4health.utils.config import load_config
+from fl4health.utils.metric_aggregation import uniform_evaluate_metrics_aggregation_fn
 
 
 def main(config: Dict[str, Any], server_checkpoint_path: Optional[Path]) -> None:
